Add TimeInterval flag to define UpdateRates period.
If timeout interval will be informed, this value will be passed to Worker Manager to schedule an UpdateRates job.
If no timeout interval defined at start, default value will be assumed. (30 minutes)
Also, in this commit we initialize Cache service and fire UpdateRates to feed it, at main app starts. ;)